// src/common/enums/file-type.enum.ts
// src/common/enums/file-type.enum.ts
export enum FileType {
  PROFILE_PICTURE = 'profile_picture',
  COMPANY_LOGO = 'company_logo',
  NGO_LOGO = 'ngo_logo',
  RESUME = 'resume',
  COVER_LETTER = 'cover_letter',
  CERTIFICATE = 'certificate',
  DOCUMENT = 'document',
  IMAGE = 'image',
  VIDEO = 'video',
  AUDIO = 'audio',
  OTHER = 'other',
}